settings/fullscreen/entity manager improvments
created a settings menu, currently only has fullscreen option

can now enter fullscreen

improvements to the entity manager, now there is a 2d array that stores an integer reference to an entity in the entity manager arraylist. The position of this integer is updated as the entity moves. The reason for this is to reduce lag when searching the entity manager arraylist, now you can find an entity on the list from a set of coordinates.